+++ src/client/usecase.ts
@@ -5,7 +5,7 @@ import { MapInterpreterError, ProfileParameterError } from '../internal';
 import { UnexpectedError } from '../internal/errors';
 import { NonPrimitive, Variables } from '../internal/interpreter/variables';
 import { Result } from '../lib';
-import { ExponentialBackoff } from '../lib/backoff';
+import { Backoff, ConstantBackoff, ExponentialBackoff } from '../lib/backoff';
 import {
   eventInterceptor,
   Events,
@@ -16,6 +16,7 @@ import {
   AbortPolicy,
   CircuitBreakerPolicy,
   FailurePolicyRouter,
+  RetryPolicy,
 } from './failure/policies';
 import { FailurePolicy, UsecaseInfo } from './failure/policy';
 import { ProfileBase } from './profile';
@@ -208,24 +209,36 @@ class UseCaseBase implements Interceptable {
 
     let policy: FailurePolicy;
     if (retryPolicyConfig.kind === OnFail.CIRCUIT_BREAKER) {
-      let backoff: ExponentialBackoff | undefined = undefined;
+      let backoff: ExponentialBackoff | undefined = new ExponentialBackoff(
+        Backoff.DEFAULT_INITIAL,
+        ExponentialBackoff.DEFAULT_BASE
+      );
       if (
         retryPolicyConfig.backoff?.kind &&
         retryPolicyConfig.backoff?.kind === BackoffKind.EXPONENTIAL
       ) {
         backoff = new ExponentialBackoff(
-          retryPolicyConfig.backoff.start ?? 2000,
-          retryPolicyConfig.backoff.factor
+          retryPolicyConfig.backoff.start ?? Backoff.DEFAULT_INITIAL,
+          retryPolicyConfig.backoff.factor ?? ExponentialBackoff.DEFAULT_BASE
         );
       }
 
       policy = new CircuitBreakerPolicy(
         usecaseInfo,
         //TODO are these defauts ok?
-        retryPolicyConfig.maxContiguousRetries ?? 5,
-        30_000,
-        retryPolicyConfig.requestTimeout,
+        retryPolicyConfig.maxContiguousRetries ??
+          RetryPolicy.DEFAULT_MAX_CONTIGUOUS_RETRIES,
+        retryPolicyConfig.openTime ?? CircuitBreakerPolicy.DEFAULT_OPEN_TIME,
+        retryPolicyConfig.requestTimeout ?? RetryPolicy.DEFAULT_REQUEST_TIMEOUT,
         backoff
+      );
+    } else if (retryPolicyConfig.kind === OnFail.SIMPLE) {
+      policy = new RetryPolicy(
+        usecaseInfo,
+        retryPolicyConfig.maxContiguousRetries ??
+          RetryPolicy.DEFAULT_MAX_CONTIGUOUS_RETRIES,
+        retryPolicyConfig.requestTimeout ?? RetryPolicy.DEFAULT_REQUEST_TIMEOUT,
+        new ConstantBackoff(0)
       );
     } else if (retryPolicyConfig.kind === OnFail.NONE) {
       policy = new AbortPolicy(usecaseInfo);