+++ src/internal/interpreter/map-interpreter.test.ts
@@ -228,6 +228,40 @@ describe('MapInterpreter', () => {
     expect(result.isOk() && result.value).toEqual(144);
   });
 
+  it('should correctly trim and parse path parameters in http call', async () => {
+    const url = '/thirteen';
+    await mockServer.get(url + '/2/get/now').thenJson(200, { data: 169 });
+
+    const interpreter = new MapInterpreter(
+      {
+        usecase: 'Test',
+        input: { page: '2', cmd: 'get', when: 'now' },
+        security: [],
+        services: mockServicesSelector,
+      },
+      { fetchInstance }
+    );
+    const ast = parseMapFromSource(`
+      map Test {
+        page = input.page
+        http GET "${url}/{ page     	  }/{input.cmd }/{ input.when}" {
+          request {
+            headers {
+              "content-type" = "application/json"
+            }
+          }
+
+          response 200 "application/json" "en-US" {
+            map result body.data
+          }
+        }
+      }`);
+    const result = await interpreter.perform(ast);
+
+    result.unwrap();
+    expect(result.isOk() && result.value).toEqual(169);
+  });
+
   it('should call an API with parameters', async () => {
     const url = '/twelve';
     await mockServer
@@ -1307,5 +1341,75 @@ describe('MapInterpreter', () => {
     );
     const result = await interpreter.perform(ast);
     expect(result.isOk() && result.value).toEqual(15);
+  });
+
+  it('should preserve buffer types with foreach', async () => {
+    const ast = parseMapFromSource(`    
+    map Test {
+      mappedItems = call foreach(item of input.items) mapItem(item = item)
+    
+      map result {
+        items: mappedItems
+      }
+    }
+    
+    operation mapItem {
+        return args.item
+    }`);
+
+    const buffer = Buffer.from('hello');
+    expect(Buffer.isBuffer(buffer)).toBe(true);
+
+    const interpreter = new MapInterpreter(
+      {
+        usecase: 'Test',
+        security: [],
+        services: ServiceSelector.withDefaultUrl(''),
+        input: {
+          items: [{ buffer }, { buffer }],
+        },
+      },
+      { fetchInstance }
+    );
+
+    const result = await interpreter.perform(ast);
+    expect(result.isOk()).toBe(true);
+    const items = (result.unwrap() as any).items;
+
+    expect(Buffer.isBuffer(items[0].buffer)).toBe(true);
+    expect(Buffer.isBuffer(items[1].buffer)).toBe(true);
+  });
+
+  it('should not leak optional properties in foreach', async () => {
+    const ast = parseMapFromSource(`    
+    map Test {
+      mappedItems = call foreach(item of input.items) mapItem(item = item)
+    
+      map result {
+        items: mappedItems
+      }
+    }
+    
+    operation mapItem {
+        return args.item
+    }`);
+
+    const interpreter = new MapInterpreter(
+      {
+        usecase: 'Test',
+        security: [],
+        services: ServiceSelector.withDefaultUrl(''),
+        input: {
+          items: [{ a: 1, b: 2 }, { a: 3 }],
+        },
+      },
+      { fetchInstance }
+    );
+
+    const result = await interpreter.perform(ast);
+    expect(result.isOk()).toBe(true);
+    const items = (result.unwrap() as any).items;
+
+    expect(items).toStrictEqual([{ a: 1, b: 2 }, { a: 3 }]);
   });
 });