+++ src/internal/interpreter/sandbox/sandbox.test.ts
@@ -114,4 +114,31 @@ describe('sandbox', () => {
     expect(v).toStrictEqual([1, 2, 3]);
     expect(Array.isArray(v)).toBe(true);
   });
+
+  it('correctly evaluates array literal inside object literal', () => {
+    const v = evalScript(
+      '{ a: 1, b: [1, 2, 3, { x: Buffer.from("hello"), y: [1, 2, 3] }] }'
+    );
+    expect(v).toStrictEqual({
+      a: 1,
+      b: [1, 2, 3, { x: Buffer.from('hello'), y: [1, 2, 3] }],
+    });
+    expect(Array.isArray((v as { b: unknown }).b)).toBe(true);
+  });
+
+  describe('stdlib', () => {
+    it('provides unstable stdlib', () => {
+      expect(
+        evalScript(
+          'std.unstable.time.isoDateToUnixTimestamp("2022-01-01T00:11:00.123Z")'
+        )
+      ).toStrictEqual(1640995860123);
+
+      expect(
+        evalScript(
+          '(std.unstable.debug.log(1123), std.unstable.time.unixTimestampToIsoDate(1640995860123))'
+        )
+      ).toStrictEqual('2022-01-01T00:11:00.123Z');
+    });
+  });
 });