+++ src/internal/interpreter/variables.ts
@@ -47,7 +47,11 @@ export function isPrimitive(input: Variables): input is Primitive {
 }
 
 export function isNonPrimitive(input: Variables): input is NonPrimitive {
-  return typeof input === 'object' && !Array.isArray(input);
+  return (
+    typeof input === 'object' &&
+    !Array.isArray(input) &&
+    !Buffer.isBuffer(input)
+  );
 }
 
 export function isEmptyRecord(
@@ -106,6 +110,17 @@ export const getValue = (
 };
 
 /**
+ * Turns a variable (both primitive and non-primitive) into a string.
+ */
+export const variableToString = (variable: Variables): string => {
+  if (typeof variable === 'string') {
+    return variable;
+  }
+
+  return JSON.stringify(variable);
+};
+
+/**
  * Stringifies a Record of variables. `undefined` values are removed.
  */
 export const variablesToStrings = (
@@ -116,7 +131,7 @@ export const variablesToStrings = (
   if (variables) {
     for (const [key, value] of Object.entries(variables)) {
       if (value !== undefined) {
-        result[key] = typeof value === 'string' ? value : JSON.stringify(value);
+        result[key] = variableToString(value);
       }
     }
   }